Preserving Rollouts + Adding Rollout Decay
Fixed problem of immediate reward in rollout.

Added tree continuity so rollouts are preserved between temporally adjacent nodes.
Added decay function so rollouts do not stick around while they are irrelevant.

Brought all of the variables we want to modulate to the beginning of main_loop.

Increased number of total rollouts for hopefully more stability. After running the code a few times, stability is unclear: the code seems to want to change medications very frequently.

This code is running for me at 100000 rollouts in 38.15 seconds.